refactor: 정적 리소스를 가져오는 코드를 Utils 클래스로 분리
- 어댑터, API, StaticHandler 등 다양한 곳에서 사용하기 위해 분리
