+++ src/main/java/codesquad/app/api/MainApi.java
@@ -1,7 +1,6 @@
 package codesquad.app.api;
 
 import codesquad.app.api.annotation.ApiMapping;
-import codesquad.http.handler.StaticHandler;
 import codesquad.http.message.SessionManager;
 import codesquad.http.message.constant.ContentType;
 import codesquad.http.message.constant.HttpMethod;
@@ -9,16 +8,18 @@ import codesquad.http.message.constant.HttpStatus;
 import codesquad.http.message.request.HttpRequest;
 import codesquad.http.message.response.HttpResponse;
 
+import static codesquad.utils.FileUtil.getStaticFiles;
+
 public class MainApi {
 
     @ApiMapping(method = HttpMethod.GET, path = "/")
     public HttpResponse main(HttpRequest request) {
         if (SessionManager.isValidSession(request.getSessionId())) {
             return HttpResponse.of(ContentType.TEXT_HTML, request.getRequestStartLine().getProtocol(),
-                    HttpStatus.OK, StaticHandler.getStaticFiles("/main/index.html"));
+                    HttpStatus.OK, getStaticFiles("/main/index.html"));
         }
 
         return HttpResponse.of(ContentType.TEXT_HTML, request.getRequestStartLine().getProtocol(),
-                HttpStatus.OK, StaticHandler.getStaticFiles("/index.html"));
+                HttpStatus.OK, getStaticFiles("/index.html"));
     }
 }
